Update liquidity table UI
